pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "pyrus-sheet-sync"
version = "0.1.0"
description = "Google Sheets to Pyrus form tasks sync"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.111.0",
    "google-auth>=2.29.0",
    "httpx>=0.27.0",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[project.scripts]
pyrus-sheet-sync = "pyrus_sheet_sync.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["pyrus_sheet_sync*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
